scraper-outstanding/requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
feedparser==6.0.11
idna==3.7
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
sgmllib3k==1.0.0
urllib3==2.2.2
